convert vegetation
