+++ src/vegetation.rs
@@ -2,7 +2,6 @@ use image::{GrayImage, Luma, Rgb, RgbImage, Rgba, RgbaImage};
 use imageproc::drawing::{draw_filled_circle_mut, draw_filled_rect_mut, draw_line_segment_mut};
 use imageproc::filter::median_filter;
 use imageproc::rect::Rect;
-use ini::Ini;
 use rustc_hash::FxHashMap as HashMap;
 use std::error::Error;
 use std::f32::consts::SQRT_2;
@@ -10,9 +9,10 @@ use std::fs::File;
 use std::io::{BufWriter, Write};
 use std::path::Path;
 
+use crate::config::Config;
 use crate::util::{read_lines, read_lines_no_alloc};
 
-pub fn makevege(thread: &String) -> Result<(), Box<dyn Error>> {
+pub fn makevege(config: &Config, thread: &String) -> Result<(), Box<dyn Error>> {
     println!("Generating vegetation...");
 
     let tmpfolder = format!("temp{}", thread);
@@ -42,13 +42,7 @@ pub fn makevege(thread: &String) -> Result<(), Box<dyn Error>> {
         }
     }
 
-    let conf = Ini::load_from_file("pullauta.ini").unwrap();
-    let block: f64 = conf
-        .general_section()
-        .get("greendetectsize")
-        .unwrap_or("3")
-        .parse::<f64>()
-        .unwrap_or(3.0);
+    let block = config.greendetectsize;
 
     let mut xyz: HashMap<(u64, u64), f64> = HashMap::default();
     let mut top: HashMap<(u64, u64), f64> = HashMap::default();
@@ -71,155 +65,31 @@ pub fn makevege(thread: &String) -> Result<(), Box<dyn Error>> {
     })
     .expect("Can not read file");
 
-    let mut zones = vec![];
-    let mut i: u32 = 1;
-    loop {
-        let last_zone = conf
-            .general_section()
-            .get(format!("zone{}", i))
-            .unwrap_or("");
-        if last_zone.is_empty() {
-            break;
-        }
-        zones.push(last_zone);
-        i += 1;
-    }
+    let zones = &config.zones;
+    let thresholds = &config.thresholds;
 
-    let thresholds = {
-        let mut thresholds = vec![];
-        let mut i: u32 = 1;
-        loop {
-            let last_threshold = conf
-                .general_section()
-                .get(format!("thresold{}", i))
-                .unwrap_or("");
-            if last_threshold.is_empty() {
-                break;
-            }
-            // parse the threshold values
-            let mut parts = last_threshold.split('|');
-            let v0: f64 = parts.next().unwrap().parse::<f64>().unwrap();
-            let v1: f64 = parts.next().unwrap().parse::<f64>().unwrap();
-            let v2: f64 = parts.next().unwrap().parse::<f64>().unwrap();
-
-            thresholds.push((v0, v1, v2));
-            i += 1;
-        }
-        thresholds
-    };
-
-    let greenshades = conf
-        .general_section()
-        .get("greenshades")
-        .unwrap_or("")
-        .split('|')
-        .map(|v| v.parse::<f64>().unwrap())
-        .collect::<Vec<f64>>();
-    let yellowheight: f64 = conf
-        .general_section()
-        .get("yellowheight")
-        .unwrap_or("0.9")
-        .parse::<f64>()
-        .unwrap_or(0.9);
-    let yellowthreshold: f64 = conf
-        .general_section()
-        .get("yellowthresold")
-        .unwrap_or("0.9")
-        .parse::<f64>()
-        .unwrap_or(0.9);
-    let greenground: f64 = conf
-        .general_section()
-        .get("greenground")
-        .unwrap_or("0.9")
-        .parse::<f64>()
-        .unwrap_or(0.9);
-    let pointvolumefactor: f64 = conf
-        .general_section()
-        .get("pointvolumefactor")
-        .unwrap_or("0.1")
-        .parse::<f64>()
-        .unwrap_or(0.1);
-    let pointvolumeexponent: f64 = conf
-        .general_section()
-        .get("pointvolumeexponent")
-        .unwrap_or("1")
-        .parse::<f64>()
-        .unwrap_or(1.0);
-    let greenhigh: f64 = conf
-        .general_section()
-        .get("greenhigh")
-        .unwrap_or("2")
-        .parse::<f64>()
-        .unwrap_or(2.0);
-    let topweight: f64 = conf
-        .general_section()
-        .get("topweight")
-        .unwrap_or("0.8")
-        .parse::<f64>()
-        .unwrap_or(0.8);
-    let greentone: f64 = conf
-        .general_section()
-        .get("lightgreentone")
-        .unwrap_or("200")
-        .parse::<f64>()
-        .unwrap_or(200.0);
-    let zoffset: f64 = conf
-        .general_section()
-        .get("vegezoffset")
-        .unwrap_or("0")
-        .parse::<f64>()
-        .unwrap_or(0.0);
-    let uglimit: f64 = conf
-        .general_section()
-        .get("undergrowth")
-        .unwrap_or("0.35")
-        .parse::<f64>()
-        .unwrap_or(0.35);
-    let uglimit2: f64 = conf
-        .general_section()
-        .get("undergrowth2")
-        .unwrap_or("0.56")
-        .parse::<f64>()
-        .unwrap_or(0.56);
-    let addition: i32 = conf
-        .general_section()
-        .get("greendotsize")
-        .unwrap_or("0")
-        .parse::<i32>()
-        .unwrap_or(0);
-    let firstandlastreturnasground = conf
-        .general_section()
-        .get("firstandlastreturnasground")
-        .unwrap_or("")
-        .parse::<u64>()
-        .unwrap_or(1);
-    let firstandlastfactor = conf
-        .general_section()
-        .get("firstandlastreturnfactor")
-        .unwrap_or("0")
-        .parse::<f64>()
-        .unwrap_or(0.0);
-    let lastfactor = conf
-        .general_section()
-        .get("lastreturnfactor")
-        .unwrap_or("0")
-        .parse::<f64>()
-        .unwrap_or(0.0);
-
-    let vege_bitmode: bool = conf.general_section().get("vege_bitmode").unwrap_or("0") == "1";
-
-    let yellowfirstlast = conf
-        .general_section()
-        .get("yellowfirstlast")
-        .unwrap_or("")
-        .parse::<u64>()
-        .unwrap_or(1);
-    let vegethin: u32 = conf
-        .general_section()
-        .get("vegethin")
-        .unwrap_or("0")
-        .parse::<u32>()
-        .unwrap_or(0);
+    let &Config {
+        vege_bitmode,
+        yellowheight,
+        yellowthreshold,
+        greenground,
+        pointvolumefactor,
+        pointvolumeexponent,
+        greenhigh,
+        topweight,
+        greentone,
+        zoffset,
+        uglimit,
+        uglimit2,
+        addition,
+        firstandlastreturnasground,
+        firstandlastfactor,
+        lastfactor,
+        yellowfirstlast,
+        vegethin,
+        ..
+    } = config;
+    let greenshades = &config.greenshades;
 
     let path = format!("{}/xyztemp.xyz", tmpfolder);
     let xyz_file_in = Path::new(&path);
@@ -387,12 +257,7 @@ pub fn makevege(thread: &String) -> Result<(), Box<dyn Error>> {
     let wy = (xmax - xmin).floor() / 3.0;
     let hy = (ymax - ymin).floor() / 3.0;
 
-    let scalefactor: f64 = conf
-        .general_section()
-        .get("scalefactor")
-        .unwrap_or("1")
-        .parse::<f64>()
-        .unwrap_or(1.0);
+    let scalefactor = config.scalefactor;
 
     let mut imgug = RgbaImage::from_pixel(
         (w * block * 600.0 / 254.0 / scalefactor) as u32,
@@ -529,29 +394,17 @@ pub fn makevege(thread: &String) -> Result<(), Box<dyn Error>> {
             }
         }
     }
-    let proceed_yellows: bool = conf
-        .general_section()
-        .get("yellow_smoothing")
-        .unwrap_or("0")
-        == "1";
-    let med: u32 = conf
-        .general_section()
-        .get("medianboxsize")
-        .unwrap_or("0")
-        .parse::<u32>()
-        .unwrap_or(0);
+
+    let proceed_yellows: bool = config.proceed_yellows;
+    let med: u32 = config.med;
+    let med2 = config.med2;
+
     if med > 0 {
         imggr1b = median_filter(&imggr1, med / 2, med / 2);
         if proceed_yellows {
             imgye2b = median_filter(&imgye2, med / 2, med / 2);
         }
     }
-    let med2: u32 = conf
-        .general_section()
-        .get("medianboxsize2")
-        .unwrap_or("0")
-        .parse::<u32>()
-        .unwrap_or(0);
     if med2 > 0 {
         imggr1 = median_filter(&imggr1b, med2 / 2, med2 / 2);
         if proceed_yellows {
@@ -639,18 +492,8 @@ pub fn makevege(thread: &String) -> Result<(), Box<dyn Error>> {
 
     let black = Rgb([0, 0, 0]);
     let blue = Rgb([29, 190, 255]);
-    let water = conf
-        .general_section()
-        .get("waterclass")
-        .unwrap_or("")
-        .parse::<u64>()
-        .unwrap_or(0);
-    let buildings = conf
-        .general_section()
-        .get("buildingsclass")
-        .unwrap_or("")
-        .parse::<u64>()
-        .unwrap_or(0);
+    let buildings = config.buildings;
+    let water = config.water;
     if buildings > 0 || water > 0 {
         read_lines_no_alloc(xyz_file_in, |line| {
             let mut parts = line.split(' ');
@@ -676,12 +519,7 @@ pub fn makevege(thread: &String) -> Result<(), Box<dyn Error>> {
         })
         .expect("Can not read file");
     }
-    let waterele = conf
-        .general_section()
-        .get("waterelevation")
-        .unwrap_or("")
-        .parse::<f64>()
-        .unwrap_or(-999999.0);
+
     let path = format!("{}/xyz2.xyz", tmpfolder);
     let xyz_file_in = Path::new(&path);
 
@@ -691,7 +529,7 @@ pub fn makevege(thread: &String) -> Result<(), Box<dyn Error>> {
         let y: f64 = parts.next().unwrap().parse::<f64>().unwrap();
         let hh: f64 = parts.next().unwrap().parse::<f64>().unwrap();
 
-        if hh < waterele {
+        if hh < config.waterele {
             draw_filled_rect_mut(
                 &mut imgwater,
                 Rect::at((x - xmin) as i32 - 1, (ymax - y) as i32 - 1).of_size(3, 3),
